feat : useInfinityProject hook
