+++ frontend/components/community/Content.tsx
@@ -9,9 +9,9 @@ import Filter from '../Filter';
 import Message from '../Message';
 import CommunityItemSkeleton from '../skeleton/CommunityItemSkeleton';
 import { ARTICLE_FILTER, POST_COMMUNITY_CATEGORY } from '@/constant/constant';
-import { getAllCommunity } from '@/util/api/getAllProject';
 import { articleFilter } from '@/util/filter/articleFilter';
 import { communityFilter } from '@/util/filter/communityFilter';
+import { getAllCommunity } from '@/util/api/getAllData';
 
 export default function Content() {
   const router = useRouter();
