+++ frontend/util/api/index.ts
@@ -3,13 +3,13 @@ import { getCookie } from '../cookie';
 
 axios.defaults.withCredentials = true;
 
-// const baseURL =
-//   process.env.NODE_ENV === 'production'
-//     ? process.env.NEXT_RESOURCE_URL
-//     : 'http://3.38.165.177:8080';
+const baseURL =
+  process.env.NODE_ENV === 'production'
+    ? process.env.NEXT_RESOURCE_URL
+    : 'http://3.38.165.177:8080';
 
 const api = axios.create({
-  baseURL: 'http://3.38.165.177:8080',
+  baseURL,
   timeout: 3000,
   withCredentials: true,
   headers: {
